requirements.txt
# This file is @generated by PDM.
# Please do not edit it manually.

annotated-types==0.7.0
anthropic==0.96.0
anyio==4.13.0
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0; python_full_version >= "3.9" and platform_python_implementation != "PyPy"
charset-normalizer==3.4.7
cryptography==46.0.7
distro==1.9.0
docstring-parser==0.18.0
dotenv==0.9.9
filetype==1.2.0
google==3.0.0
google-analytics-data==0.21.0
google-api-core[grpc]==2.30.3
google-api-python-client==2.194.0
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1
google-auth[requests]==2.49.2
google-genai==1.73.1
googleapis-common-protos==1.74.0
greenlet==3.4.0
grpcio==1.80.0
grpcio-status==1.80.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
jiter==0.14.0
jsonpatch==1.33
jsonpointer==3.1.1
langchain==1.2.15
langchain-anthropic==1.4.0
langchain-core==1.2.28
langchain-google-genai==4.2.1
langgraph==1.1.6
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.13
langsmith==0.7.30
oauthlib==3.3.1
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.0
playwright==1.58.0
proto-plus==1.27.2
protobuf==6.33.6
pyasn1==0.6.3
pyasn1-modules==0.4.2
pycparser==3.0; python_full_version >= "3.9" and platform_python_implementation != "PyPy" and implementation_name != "PyPy"
pydantic==2.13.0
pydantic-core==2.46.0
pyee==13.0.1
pyparsing==3.3.2
python-dotenv==1.2.2
pyyaml==6.0.3
requests==2.33.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
sniffio==1.3.1
soupsieve==2.8.3
steel-sdk @ git+https://github.com/steel-dev/steel-python@541d5c7cc9a2206f181784f491c7c02fc26cc331
tenacity==9.1.4
typing-extensions==4.15.0
typing-inspection==0.4.2
uritemplate==4.2.0
urllib3==2.6.3
uuid-utils==0.14.1
websockets==16.0
xxhash==3.6.0
zstandard==0.25.0
